requirements.txt
slack-bolt
slack-sdk
google-genai
python-dotenv
